backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.34
asyncpg==0.29.0
psycopg[binary]==3.2.3
alembic==1.13.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.2
pydantic-settings==2.6.0
python-multipart==0.0.12
httpx==0.27.2
redis==5.0.8
aiosqlite==0.20.0
tenacity==9.0.0
loguru==0.7.2
stripe==6.7.0

